pyproject.toml
[tool.poetry]
name = "procollab-skills"
version = "0.1.0"
description = "prototype"
authors = ["Alexey Kudelko <flexivanov237@gmail.com>", "Igor Kuzmenkov <tkdomigor@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 120

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.8.0"
attrs = "23.2.0"
black = "24.3.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
django-filter = "24.1"
django = "5.0.3"
djangorestframework = "3.15.0"
drf-spectacular-sidecar = "2024.3.4"
drf-spectacular = "0.27.1"
idna = "3.6"
image = "1.5.33"
inflection = "0.5.1"
jsonschema-specifications = "2023.12.1"
jsonschema = "4.21.1"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
packaging = "24.0"
pathspec = "0.12.1"
pillow = "10.2.0"
pip = "24.0"
platformdirs = "4.2.0"
pycparser = "2.21"
pyyaml = "6.0.1"
referencing = "0.34.0"
requests = "2.31.0"
rpds-py = "0.18.0"
setuptools = "69.0.3"
six = "1.16.0"
sqlparse = "0.4.4"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
webp = "0.3.0"
wheel = "0.42.0"
python-decouple = "^3.8"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
yookassa = "^3.1.0"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.3.1"
djangorestframework-dataclasses = "^1.3.1"
djangorestframework-simplejwt = "^5.3.1"
pyright = "^1.1.361"
django-stubs = "^5.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
